+++ src/imio/annex/browser/views.py
@@ -1,0 +1,66 @@
+# -*- coding: utf-8 -*-
+
+from collective.eeafaceted.batchactions import _ as _CEBA
+from collective.eeafaceted.batchactions.browser.views import BaseBatchActionForm
+from imio.annex import _
+from imio.annex.content.annex import IAnnex
+from io import BytesIO
+from plone import api
+from plone.rfc822.interfaces import IPrimaryFieldInfo
+
+import zipfile
+
+
+class DownloadAnnexesBatchActionForm(BaseBatchActionForm):
+
+    label = _CEBA("Download annexes")
+    button_with_icon = False
+    section = "annexes"
+
+    def available(self):
+        """ """
+        return True
+
+    def zipfiles(self, content):
+        """Return the path and file stream of all content we find here"""
+        fstream = BytesIO()
+        zipper = zipfile.ZipFile(fstream, 'w', zipfile.ZIP_DEFLATED)
+        for obj in content:
+            if not IAnnex.providedBy(obj):
+                continue
+            primary_field = IPrimaryFieldInfo(obj)
+            data = primary_field.value.data
+            filename = primary_field.value.filename
+            zipper.writestr(filename, data)
+            created = obj.created()
+            zipper.NameToInfo[filename].date_time = (
+                created.year(), created.month(), created.day(), created.hour(), created.minute(),
+                int(created.second()))
+        zipper.close()
+        return fstream.getvalue()
+
+    def _apply(self, **data):
+        """ """
+        try:
+            return self.do_zip()
+        except zipfile.LargeZipFile:
+            message = _("Too much annexes to zip, try selecting fewer annexes...")
+            api.portal.show_message(message, self.request, type="error")
+            return self.request.response.redirect(self.context.absolute_url())
+
+    def do_zip(self):
+        """ Zip all of the content in this location (context)"""
+        self.request.response.setHeader('Content-Type', 'application/zip')
+        self.request.response.setHeader('Content-disposition', 'attachment;filename=%s.zip'
+                                        % self.context.getId())
+        content = [brain.getObject() for brain in self.brains]
+        zipped_content = self.zipfiles(content)
+        self.request.set('zip_file_content', zipped_content)
+        return zipped_content
+
+    def render(self):
+        """ """
+        if 'zip_file_content' in self.request:
+            return self.request['zip_file_content']
+        else:
+            return super(DownloadAnnexesBatchActionForm, self).render()